requirement.txt
absl-py==0.11.0
alabaster==0.7.11
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-live==3.9.7
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.0.4
astropy==3.0.4
astunparse==1.6.3
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
baidu-aip==2.2.18.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
bokeh==0.13.0
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.2.1
certifi==2018.8.24
cffi==1.14.5
chardet==4.0.0
click==6.7
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.7
conda==4.5.11
conda-build==3.15.1
constantly==15.1.0
contextlib2==0.5.5
cos-python-sdk-v5==1.9.6
crcmod==1.7
cryptography==3.4.7
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
dataprocessor==0.0.11
datashape==0.5.4
DateTime==4.3
decorator==4.3.0
defusedxml==0.5.0
dicttoxml==1.7.4
distributed==1.23.1
Django==2.2.2
django-cors-headers==3.7.0
django-extensions==3.1.2
django-fsm==2.7.1
djangorestframework==3.12.3
djangorestframework-jwt==1.11.0
docker==5.0.0
docutils==0.14
english==2020.7.0
entrypoints==0.2.3
et-xmlfile==1.0.1
Faker==6.5.0
fastcache==1.0.2
filelock==3.0.8
Flask==1.0.2
Flask-Cors==3.0.6
flatbuffers==1.12
gast==0.3.3
gettext-windows==1.0
gevent==1.3.6
glob2==0.6
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.32.0
h5py==2.10.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.10
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==3.7.0
incremental==17.5.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jmespath==0.10.0
jsonpatch==1.28
jsonpointer==2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown==3.3.4
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
mock==4.0.3
monotonic==1.5
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
mysqlclient==2.0.3
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
neotime==1.7.4
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.19.5
numpydoc==0.8.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.46
opencv-python==4.5.1.48
openpyxl==2.5.6
opt-einsum==3.3.0
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
pansi==2020.7.3
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib==1.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.2.0
pkginfo==1.4.2
pluggy==0.7.1
ply==3.11
prometheus-client==0.3.1
prompt-toolkit==2.0.10
protobuf==3.15.3
psutil==5.4.7
py==1.6.0
py2neo==2021.0.1
pyahocorasick==1.4.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.10.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.2.0
PyJWT==1.7.1
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-gettext==4.0
pytz==2018.5
PyWavelets==1.0.0
pywin32==227
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qcloud==0.0.1
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.11.0
rsa==4.7.2
ruamel-yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simpleui==4.0.2
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
sqlparse==0.4.1
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.4
tblib==1.3.2
tencentcloud-sdk-python==3.0.418
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.3
toolz==0.9.0
torch==1.7.1
torchfile==0.1.0
torchvision==0.4.1
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
Twisted==18.7.0
typing==3.7.4.3
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.4
visdom==0.1.8.9
vod-python-sdk==1.4.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.14.1
widgetsnbextension==3.4.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zipp==3.4.0
zope.interface==4.5.0
